process review comments; move toTimestampStringMilli() to Logging namespace
